only use one port, and use "real name" to choose the connection
See examples/config.yml, the config format has changed. There is now only one
server, listening on a single port, that everybody connects to for all
connections and clients.

Choosing the connection and client is done by setting the "real name" to a
special format outlined in the config. At first I was doing this with a
`/tkellem connect <conn> <client>` command, but it turns out there's some
problems with how IRC clients handle that. I might revisit this later.

Eventually this will need to look at the password as well.