password authentication
The config.yml can now take an optional password for each server connection,
which will be checked during the connect process. The password in the .yml is
the sha1 digest, rather than the plain text of the password.

This is actually implemented as a more generic callback, so other auth methods
can be easily used (we'll probably be using LDAP).

Hmm, possibly we'll need to make the callback async.