switch to Celluloid and Sequel
This is sort of the bare minimum conversion, it doesn't really take
proper advantage of Celluloid yet. There's probably stuff I've missed
too.

The switch to Sequel was required because otherwise I'd have to wrap
every single ActiveRecord call in
ActiveRecord::Base.with_connection { ... } in order to avoid leaking
connections and eventually maxing out the connection pool. Sequel
immediately returns the connection to the pool when it's done with the
transaction, rather than holding on to it until the thread explicitly
releases it.